requirements.txt
requests
pyhamcrest
pytest
selene~=2.0.0b17
